# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api gestion planning simplon",
        "description": "Consommer les données générées par le back, pour afficher les examens prévus pour les etudiants",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/jury/exam/{token}": {
            "get": {
                "tags": [
                    "token"
                ],
                "summary": "Get all exams with activities and all students with their hour for a jury of a promotion",
                "description": "Returns all exams with activities and all students with their hour for a jury of a promotion. Token is given for an exam that is linked to a promotion",
                "operationId": "GetListForJury",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/user/exam?token={token}": {
            "get": {
                "tags": [
                    "token"
                ],
                "summary": "Get all exams for a user",
                "description": "Returns all exams with there activities, hours and date for a specific student",
                "operationId": "GetAllExamsByUser",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}